medical-records/requirements.txt
azure-storage-blob
azure-functions